requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0 # For reading .env file
pydantic>=2.0.0
numpy>=1.20.0
pandas>=2.0.0
comtypes>=1.1.14  # For Aspen Plus automation
pytest>=7.0.0